feat(shopping): move shopping-list add flow to its own page
Alternative to #251's popup-embedded UI, which felt cramped once
search results filled the popup's capped 560px height. Moves the
search -> pick recipe -> pick list -> add flow to a dedicated page
(entrypoints/shopping/), mirroring the existing Activity Log page
pattern. The popup now just has a "Add Recipe to Shopping List"
link that opens it in a new tab, same as "Open Activity Log".

No new permissions or manifest changes.
