+++ entrypoints/shopping/main.tsx
@@ -1,0 +1,12 @@
+import './style.css';
+
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+
+import { ShoppingPage } from './ShoppingPage';
+
+createRoot(document.getElementById('root')!).render(
+    <StrictMode>
+        <ShoppingPage />
+    </StrictMode>,
+);
