+++ entrypoints/shopping/style.css
@@ -1,0 +1,182 @@
+* {
+    box-sizing: border-box;
+}
+
+body {
+    margin: 0;
+    font-family:
+        -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
+    background: #1a1a1a;
+    color: #e0e0e0;
+}
+
+.shopping-page {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+    max-width: 640px;
+    margin: 0 auto;
+    padding: 1.5rem;
+}
+
+.shopping-page-header {
+    margin-bottom: 1rem;
+}
+
+.shopping-page-header h1 {
+    margin: 0;
+    font-size: 1.5rem;
+    color: #ec7e19;
+}
+
+.shopping-page-empty {
+    color: #888;
+    text-align: center;
+    padding: 2rem;
+}
+
+.shopping-search-row {
+    display: flex;
+    gap: 0.5rem;
+    align-items: stretch;
+}
+
+.shopping-search-row input[type='text'] {
+    flex: 1;
+    padding: 0.6rem 0.7rem;
+    border-radius: 4px;
+    border: 1px solid #444;
+    background-color: #2c2c2c;
+    color: #e0e0e0;
+    font-size: 1rem;
+}
+
+.shopping-search-row button {
+    white-space: nowrap;
+    padding: 0.6rem 1rem;
+    background: #ec7e19;
+    border: none;
+    border-radius: 4px;
+    color: white;
+    cursor: pointer;
+    font-size: 0.95rem;
+}
+
+.shopping-search-row button:hover:not(:disabled) {
+    background: #d66f15;
+}
+
+.shopping-search-row button:disabled {
+    background: #a65726;
+    color: rgba(255, 255, 255, 0.8);
+    cursor: not-allowed;
+    opacity: 0.7;
+}
+
+.shopping-results {
+    list-style: none;
+    margin: 1rem 0 0;
+    padding: 0;
+    display: flex;
+    flex-direction: column;
+    max-height: 320px;
+    overflow-y: auto;
+    border: 1px solid #333;
+    border-radius: 6px;
+}
+
+.shopping-results li:not(:first-child) {
+    border-top: 1px solid #333;
+}
+
+.shopping-results button {
+    width: 100%;
+    text-align: left;
+    font-size: 0.95rem;
+    padding: 0.7rem 0.9rem;
+    background-color: #222;
+    color: #e0e0e0;
+    border: none;
+    cursor: pointer;
+}
+
+.shopping-results button:hover {
+    background-color: rgba(236, 126, 25, 0.18);
+}
+
+.shopping-selected {
+    margin-top: 1rem;
+    padding: 1rem;
+    background-color: #0d0d0d;
+    border: 1px solid #333;
+    border-radius: 6px;
+}
+
+.shopping-selected p {
+    margin-top: 0;
+}
+
+.shopping-selected select {
+    width: 100%;
+    padding: 0.6rem;
+    margin-bottom: 0.6rem;
+    border-radius: 4px;
+    border: 1px solid #444;
+    background-color: #2c2c2c;
+    color: #e0e0e0;
+    font-size: 0.95rem;
+}
+
+.shopping-selected select + button {
+    width: 100%;
+    padding: 0.6rem;
+    background: #ec7e19;
+    border: none;
+    border-radius: 4px;
+    color: white;
+    cursor: pointer;
+    font-size: 0.95rem;
+}
+
+.shopping-selected select + button:hover:not(:disabled) {
+    background: #d66f15;
+}
+
+.shopping-selected select + button:disabled {
+    background: #a65726;
+    cursor: not-allowed;
+    opacity: 0.7;
+}
+
+.shopping-back {
+    display: block;
+    margin-top: 0.6rem;
+    background: transparent;
+    border: none;
+    color: #ec7e19;
+    cursor: pointer;
+    padding: 0.3rem 0;
+    font-size: 0.9rem;
+}
+
+.shopping-back:hover {
+    text-decoration: underline;
+}
+
+.shopping-empty {
+    font-size: 0.9rem;
+    color: #888;
+    margin: 1rem 0 0;
+}
+
+.shopping-error {
+    font-size: 0.9rem;
+    color: #f0c4a8;
+    margin: 1rem 0 0;
+}
+
+.shopping-success {
+    font-size: 0.9rem;
+    color: #8fd694;
+    margin: 1rem 0 0;
+}
